style: clear trending code
